fix: resolve errors and lint
