+++ semantic_router/index/base.py
@@ -1,4 +1,5 @@
 from typing import Any, List, Optional, Tuple, Union, Dict
+import json
 
 import numpy as np
 from pydantic.v1 import BaseModel
@@ -38,6 +39,20 @@ class BaseIndex(BaseModel):
         """
         raise NotImplementedError("This method should be implemented by subclasses.")
 
+    def get_utterances(self) -> List[Tuple]:
+        """Gets a list of route and utterance objects currently stored in the
+        index, including additional metadata.
+
+        :return: A list of tuples, each containing route, utterance, function
+        schema and additional metadata.
+        :rtype: List[Tuple]
+        """
+        _, metadata = self._get_all(include_metadata=True)
+        route_tuples: List[
+            Tuple[str, str, Optional[Dict[str, Any]], Dict[str, Any]]
+        ] = [(x["sr_route"], x["sr_utterance"], None, {}) for x in metadata]
+        return route_tuples
+
     def get_routes(self) -> List[Route]:
         """Gets a list of route objects currently stored in the index.
 
@@ -45,25 +60,24 @@ class BaseIndex(BaseModel):
         :rtype: List[Route]
         """
         route_tuples = self.get_utterances()
-        routes_dict: Dict[str, List[str]] = {}
-        # first create a dictionary of routes mapping to all their utterances,
-        # function_schema, and metadata
+        routes_dict: Dict[str, Route] = {}
+        # first create a dictionary of route names to Route objects
         for route_name, utterance, function_schema, metadata in route_tuples:
-            routes_dict.setdefault(
-                route_name,
-                {
-                    "function_schemas": None,
-                    "metadata": {},
-                },
-            )
-            routes_dict[route_name]["utterances"] = routes_dict[route_name].get(
-                "utterances", []
-            )
-            routes_dict[route_name]["utterances"].append(utterance)
+            # if the route is not in the dictionary, add it
+            if route_name not in routes_dict:
+                routes_dict[route_name] = Route(
+                    name=route_name,
+                    utterances=[utterance],
+                    function_schemas=function_schema,
+                    metadata=metadata,
+                )
+            else:
+                # otherwise, add the utterance to the route
+                routes_dict[route_name].utterances.append(utterance)
         # then create a list of routes from the dictionary
         routes: List[Route] = []
-        for route_name, route_data in routes_dict.items():
-            routes.append(Route(name=route_name, **route_data))
+        for route_name, route in routes_dict.items():
+            routes.append(route)
         return routes
 
     def _remove_and_sync(self, routes_to_delete: dict):
@@ -181,5 +195,77 @@ class BaseIndex(BaseModel):
         """
         raise NotImplementedError("This method should be implemented by subclasses.")
 
+    def _get_all(self, prefix: Optional[str] = None, include_metadata: bool = False):
+        """
+        Retrieves all vector IDs from the index.
+
+        This method should be implemented by subclasses.
+
+        :param prefix: The prefix to filter the vectors by.
+        :type prefix: Optional[str]
+        :param include_metadata: Whether to include metadata in the response.
+        :type include_metadata: bool
+        :return: A tuple containing a list of vector IDs and a list of metadata dictionaries.
+        :rtype: tuple[list[str], list[dict]]
+        """
+        raise NotImplementedError("This method should be implemented by subclasses.")
+
+    async def _async_get_all(
+        self, prefix: Optional[str] = None, include_metadata: bool = False
+    ) -> tuple[list[str], list[dict]]:
+        """Retrieves all vector IDs from the index asynchronously.
+
+        This method should be implemented by subclasses.
+
+        :param prefix: The prefix to filter the vectors by.
+        :type prefix: Optional[str]
+        :param include_metadata: Whether to include metadata in the response.
+        :type include_metadata: bool
+        :return: A tuple containing a list of vector IDs and a list of metadata dictionaries.
+        :rtype: tuple[list[str], list[dict]]
+        """
+        raise NotImplementedError("This method should be implemented by subclasses.")
+
+    async def _async_get_routes(self) -> List[Tuple]:
+        """Asynchronously gets a list of route and utterance objects currently
+        stored in the index, including additional metadata.
+
+        :return: A list of tuples, each containing route, utterance, function
+        schema and additional metadata.
+        :rtype: List[Tuple]
+        """
+        _, metadata = await self._async_get_all(include_metadata=True)
+        route_info = parse_route_info(metadata=metadata)
+        return route_info  # type: ignore
+
     class Config:
         arbitrary_types_allowed = True
+
+
+def parse_route_info(metadata: List[Dict[str, Any]]) -> List[Tuple]:
+    """Parses metadata from index to extract route, utterance, function
+    schema and additional metadata.
+
+    :param metadata: List of metadata dictionaries.
+    :type metadata: List[Dict[str, Any]]
+    :return: A list of tuples, each containing route, utterance, function schema and additional metadata.
+    :rtype: List[Tuple]
+    """
+    route_info = []
+    for record in metadata:
+        sr_route = record.get("sr_route", "")
+        sr_utterance = record.get("sr_utterance", "")
+        sr_function_schema = json.loads(record.get("sr_function_schema", "{}"))
+        if sr_function_schema == {}:
+            sr_function_schema = None
+
+        additional_metadata = {
+            key: value
+            for key, value in record.items()
+            if key not in ["sr_route", "sr_utterance", "sr_function_schema"]
+        }
+        # TODO: Not a fan of tuple packing here
+        route_info.append(
+            (sr_route, sr_utterance, sr_function_schema, additional_metadata)
+        )
+    return route_info
